Extract FileTree and DiffPane clusters from changes-tab.tsx (940 → 359 lines)
Splits the oversized changes-tab.tsx into three focused files:
- changes-file-tree.tsx (231 lines): FileTree sidebar with tree building/collapsing logic
- changes-diff-section.tsx (371 lines): DiffPane, FileDiffSection, and FileDiffContent
- changes-tab.tsx (358 lines): ChangesTab orchestrator with state management
