+++ apps/web/src/components/app/changes-file-tree.tsx
@@ -1,0 +1,231 @@
+import { useMemo } from "react";
+import {
+  ChevronDown,
+  ChevronRight,
+  FileCode2,
+  FileDiff,
+  FileMinus,
+  FilePlus,
+  PanelLeftClose,
+  PanelLeftOpen,
+} from "lucide-react";
+import { AnimatePresence, motion } from "framer-motion";
+
+import type { DiffFile } from "@/hooks/use-agent-diff";
+import { cn } from "@/lib/utils";
+
+type TreeNode = {
+  name: string;
+  path: string;
+  file?: DiffFile;
+  children: Map<string, TreeNode>;
+};
+
+function buildTree(files: DiffFile[]): TreeNode {
+  const root: TreeNode = { name: "", path: "", children: new Map() };
+  for (const file of files) {
+    const parts = file.path.split("/");
+    let node = root;
+    for (let i = 0; i < parts.length; i++) {
+      const part = parts[i]!;
+      const partPath = parts.slice(0, i + 1).join("/");
+      if (!node.children.has(part)) {
+        node.children.set(part, {
+          name: part,
+          path: partPath,
+          children: new Map(),
+        });
+      }
+      node = node.children.get(part)!;
+    }
+    node.file = file;
+  }
+  return collapseTree(root);
+}
+
+function collapseTree(node: TreeNode): TreeNode {
+  if (node.children.size === 1 && !node.file) {
+    const child = [...node.children.values()][0]!;
+    if (child.file && child.children.size === 0) {
+      const newChildren = new Map<string, TreeNode>();
+      newChildren.set(child.name, child);
+      return { ...node, children: newChildren };
+    }
+    const collapsed: TreeNode = {
+      name: node.name ? `${node.name}/${child.name}` : child.name,
+      path: child.path,
+      file: child.file,
+      children: child.children,
+    };
+    return collapseTree(collapsed);
+  }
+  const newChildren = new Map<string, TreeNode>();
+  for (const [key, child] of node.children) {
+    newChildren.set(key, collapseTree(child));
+  }
+  return { ...node, children: newChildren };
+}
+
+function treeFileIcon(file: DiffFile): JSX.Element {
+  if (file.status === "renamed")
+    return <FileCode2 className="h-3.5 w-3.5 text-purple-400" />;
+  if (file.status === "deleted" || (file.deleted > 0 && file.added === 0))
+    return <FileMinus className="h-3.5 w-3.5 text-status-blocked" />;
+  if (file.status === "added" || (file.added > 0 && file.deleted === 0))
+    return <FilePlus className="h-3.5 w-3.5 text-status-working" />;
+  return <FileDiff className="h-3.5 w-3.5 text-muted-foreground" />;
+}
+
+type FileTreeProps = {
+  files: DiffFile[];
+  selectedFile: string | null;
+  onSelectFile: (path: string) => void;
+  open: boolean;
+  onToggleOpen: () => void;
+  collapsedDirs: Set<string>;
+  onToggleDir: (path: string) => void;
+};
+
+export function FileTree({
+  files,
+  selectedFile,
+  onSelectFile,
+  open,
+  onToggleOpen,
+  collapsedDirs,
+  onToggleDir,
+}: FileTreeProps): JSX.Element {
+  const tree = useMemo(() => buildTree(files), [files]);
+
+  return (
+    <div
+      className="flex shrink-0 flex-col border-r border-border/50 bg-muted/20 overflow-hidden transition-[width] duration-200 ease-in-out"
+      style={{ width: open ? "14rem" : "2.25rem" }}
+    >
+      <div className="shrink-0 flex items-center border-b border-border/40 bg-muted/30">
+        {open && (
+          <span className="flex-1 px-3 py-2 text-[10px] font-semibold uppercase tracking-wider text-muted-foreground whitespace-nowrap">
+            {files.length} file{files.length !== 1 ? "s" : ""} changed
+          </span>
+        )}
+        <button
+          type="button"
+          onClick={onToggleOpen}
+          className={cn(
+            "shrink-0 flex items-center justify-center text-muted-foreground hover:bg-muted/50 cursor-pointer",
+            open ? "p-2" : "flex-1 py-2"
+          )}
+        >
+          {open ? (
+            <PanelLeftClose className="h-3.5 w-3.5" />
+          ) : (
+            <PanelLeftOpen className="h-3.5 w-3.5" />
+          )}
+        </button>
+      </div>
+      {open && (
+        <div className="min-h-0 flex-1 overflow-y-auto py-1">
+          {(tree.file ? [tree] : [...tree.children.values()]).map((child) => (
+            <TreeEntry
+              key={child.path}
+              node={child}
+              depth={0}
+              selectedFile={selectedFile}
+              onSelectFile={onSelectFile}
+              collapsedDirs={collapsedDirs}
+              onToggleDir={onToggleDir}
+            />
+          ))}
+        </div>
+      )}
+    </div>
+  );
+}
+
+type TreeEntryProps = {
+  node: TreeNode;
+  depth: number;
+  selectedFile: string | null;
+  onSelectFile: (path: string) => void;
+  collapsedDirs: Set<string>;
+  onToggleDir: (path: string) => void;
+};
+
+function TreeEntry({
+  node,
+  depth,
+  selectedFile,
+  onSelectFile,
+  collapsedDirs,
+  onToggleDir,
+}: TreeEntryProps): JSX.Element {
+  const isDir = node.children.size > 0 && !node.file;
+  const isCollapsed = collapsedDirs.has(node.path);
+  const isSelected = node.file && selectedFile === node.file.path;
+
+  const indent = depth * 16 + 8;
+
+  if (isDir) {
+    return (
+      <>
+        <button
+          type="button"
+          className="flex w-full items-center gap-1.5 px-2 py-1 text-left text-xs text-muted-foreground hover:bg-muted/40"
+          style={{ paddingLeft: `${indent}px` }}
+          title={node.name}
+          onClick={() => onToggleDir(node.path)}
+        >
+          {isCollapsed ? (
+            <ChevronRight className="h-3 w-3 shrink-0" />
+          ) : (
+            <ChevronDown className="h-3 w-3 shrink-0" />
+          )}
+          <span className="truncate">{node.name}</span>
+        </button>
+        <AnimatePresence initial={false}>
+          {!isCollapsed && (
+            <motion.div
+              key="children"
+              initial={{ height: 0, opacity: 0 }}
+              animate={{ height: "auto", opacity: 1 }}
+              exit={{ height: 0, opacity: 0 }}
+              transition={{ duration: 0.15, ease: "easeInOut" }}
+              className="overflow-hidden border-l border-border"
+              style={{ marginLeft: `${indent + 7}px` }}
+            >
+              {[...node.children.values()].map((child) => (
+                <TreeEntry
+                  key={child.path}
+                  node={child}
+                  depth={0}
+                  selectedFile={selectedFile}
+                  onSelectFile={onSelectFile}
+                  collapsedDirs={collapsedDirs}
+                  onToggleDir={onToggleDir}
+                />
+              ))}
+            </motion.div>
+          )}
+        </AnimatePresence>
+      </>
+    );
+  }
+
+  if (!node.file) return <></>;
+
+  return (
+    <button
+      type="button"
+      className={cn(
+        "flex w-full items-center gap-1.5 px-2 py-1 text-left text-xs hover:bg-muted/40",
+        isSelected && "bg-muted/60 text-foreground"
+      )}
+      style={{ paddingLeft: `${indent}px` }}
+      title={node.file.path}
+      onClick={() => onSelectFile(node.file!.path)}
+    >
+      {treeFileIcon(node.file)}
+      <span className="min-w-0 flex-1 truncate">{node.name}</span>
+    </button>
+  );
+}
